requirements.txt
fastapi ~= 0.116.1
aiohttp ~= 3.10.2
Pillow~=10.1.0
